requirements.txt
# Core AI
anthropic>=0.40.0

# Web Search
tavily-python>=0.5.0

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Async & Concurrency
anyio>=4.4.0
asyncio-throttle>=1.0.2

# Data & Serialization
orjson>=3.10.0

# Logging & Observability
structlog>=24.2.0
rich>=13.7.0

# Storage / Memory
diskcache>=5.6.3

# Utilities
tenacity>=8.3.0
httpx>=0.27.0
python-dateutil>=2.9.0

# Dev / Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
ruff>=0.4.0
mypy>=1.10.0
